Change nat extraction to uint64_t and fix string.length() sign conversion
- NatIntStd.v: extract nat as uint64_t (was unsigned int); update all
  loop counters, conversion functions, and numeral format to UINT64_C(%n)
- Std.v: cast PrimString.length() result to int64_t to avoid sign-conversion
  warning (size_t → int64_t)
- ZInt.v: fix Z.to_nat cast from unsigned int to uint64_t to match new nat type

Update all generated test golden files (.h/.cpp) and hand-written test
files (.t.cpp) to use uint64_t instead of unsigned int for nat values.
Fix loopify_itree_reified.v inline override to use uint64_t/UINT64_C.
Fix z_overflow.t.cpp: big_nat now correctly holds 4294967296 (no wrapping).

All 577 non-wip tests pass; 7 known wip tests still fail.

+++ tests/basics/ack/ack.cpp
@@ -1,20 +1,20 @@
 #include "ack.h"
 
-unsigned int Ack::ack(unsigned int m, unsigned int n) {
-  auto ack_m_impl = [&](auto &_self_ack_m, unsigned int n0) -> unsigned int {
+uint64_t Ack::ack(uint64_t m, uint64_t n) {
+  auto ack_m_impl = [&](auto &_self_ack_m, uint64_t n0) -> uint64_t {
     if (m <= 0) {
       return (n0 + 1);
     } else {
-      unsigned int pm = m - 1;
+      uint64_t pm = m - 1;
       if (n0 <= 0) {
         return ack(pm, Nat::one);
       } else {
-        unsigned int pn = n0 - 1;
+        uint64_t pn = n0 - 1;
         return ack(pm, _self_ack_m(_self_ack_m, pn));
       }
     }
   };
-  auto ack_m = [&](unsigned int n0) -> unsigned int {
+  auto ack_m = [&](uint64_t n0) -> uint64_t {
     return ack_m_impl(ack_m_impl, n0);
   };
   return ack_m(n);
