+++ tests/basics/record_match/record_match.h
@@ -3,15 +3,15 @@
 
 struct RecordMatch {
   struct MyRec {
-    unsigned int f1;
-    unsigned int f2;
-    unsigned int f3;
+    uint64_t f1;
+    uint64_t f2;
+    uint64_t f3;
 
     // ACCESSORS
     MyRec clone() const { return MyRec{(*this).f1, (*this).f2, (*this).f3}; }
   };
 
-  static unsigned int sum(const MyRec &r);
+  static uint64_t sum(const MyRec &r);
 };
 
 #endif // INCLUDED_RECORD_MATCH
