+++ tests/basics/sort/sort.t.cpp
@@ -38,16 +38,16 @@ void aSsErT(bool condition, const char *message, int line) {
 #define ASSERT(X) aSsErT(!(X), #X, __LINE__);
 
 // Helper to convert list to vector for testing
-std::vector<unsigned int>
-list_to_vector(const List<unsigned int> &l) {
-  std::vector<unsigned int> result;
-  const List<unsigned int> *current = &l;
+std::vector<uint64_t>
+list_to_vector(const List<uint64_t> &l) {
+  std::vector<uint64_t> result;
+  const List<uint64_t> *current = &l;
   while (true) {
     bool done = false;
-    std::visit(Overloaded{[&](const typename List<unsigned int>::Nil &) {
+    std::visit(Overloaded{[&](const typename List<uint64_t>::Nil &) {
                             done = true;
                           },
-                          [&](const typename List<unsigned int>::Cons &c) {
+                          [&](const typename List<uint64_t>::Cons &c) {
                             result.push_back(c.a0);
                             current = c.a1.get();
                           }},
@@ -59,11 +59,11 @@ list_to_vector(const List<unsigned int> &l) {
 }
 
 // Helper to create a list from a vector
-List<unsigned int>
-vector_to_list(const std::vector<unsigned int> &vec) {
-  auto result = List<unsigned int>::nil();
+List<uint64_t>
+vector_to_list(const std::vector<uint64_t> &vec) {
+  auto result = List<uint64_t>::nil();
   for (auto it = vec.rbegin(); it != vec.rend(); ++it) {
-    result = List<unsigned int>::cons(*it, result);
+    result = List<uint64_t>::cons(*it, result);
   }
   return result;
 }
@@ -75,7 +75,7 @@ void test_sort(const std::string &name, SortFn sortFn) {
 
   // Empty list
   {
-    auto empty = List<unsigned int>::nil();
+    auto empty = List<uint64_t>::nil();
     auto result = sortFn(empty);
     auto sorted_list = result.x;
     auto vec = list_to_vector(sorted_list);
@@ -85,7 +85,7 @@ void test_sort(const std::string &name, SortFn sortFn) {
   // Single element
   {
     auto single =
-        List<unsigned int>::cons(5, List<unsigned int>::nil());
+        List<uint64_t>::cons(5, List<uint64_t>::nil());
     auto result = sortFn(single);
     auto sorted_list = result.x;
     auto vec = list_to_vector(sorted_list);
