+++ tests/basics/string_match/string_match.cpp
@@ -1,3 +1,5 @@
 #include "string_match.h"
 
-bool StringMatch::is_empty(std::string s) { return s.length() == int64_t(0); }
+bool StringMatch::is_empty(std::string s) {
+  return static_cast<int64_t>(s.length()) == int64_t(0);
+}
