+++ tests/regression/sep_ext_name_collision/SetA.h
@@ -3,7 +3,7 @@
 
 namespace SetA {
 
-unsigned int make(unsigned int n);
+uint64_t make(uint64_t n);
 
 } // namespace SetA
